feat: separate tx_script & add compile_tx_script
