+++ crates/miden-objects/src/transaction/mod.rs
@@ -14,6 +14,7 @@ mod partial_blockchain;
 mod proven_tx;
 mod transaction_id;
 mod tx_args;
+mod tx_script;
 mod tx_header;
 mod tx_witness;
 
@@ -26,6 +27,7 @@ pub use proven_tx::{
     InputNoteCommitment, ProvenTransaction, ProvenTransactionBuilder, TxAccountUpdate,
 };
 pub use transaction_id::TransactionId;
-pub use tx_args::{TransactionArgs, TransactionScript};
+pub use tx_args::TransactionArgs;
+pub use tx_script::TransactionScript;
 pub use tx_header::TransactionHeader;
 pub use tx_witness::TransactionWitness;
